mqtt-alarm-extractor/requirements.txt
cognite-sdk>=7.0.0
paho-mqtt>=2.0.0
pyyaml>=6.0

